Rename files in amaranth_ext
