+++ transactron/utils/amaranth_ext/__init__.py
@@ -1,2 +1,2 @@
-from .amaranth_ext import *  # noqa: F401
-from .modules import *  # noqa: F401
+from .functions import *  # noqa: F401
+from .elaboratables import *  # noqa: F401
