+++ transactron/utils/amaranth_ext/functions.py
